garvis-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "garvis-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "faker",
  "numpy",
  "pandas",
  "duckdb",
  "uv",
  "google-cloud-speech",
  "google-cloud-texttospeech",
  "janus",
  "langgraph",
  "langgraph-checkpoint-sqlite",
  "langgraph-checkpoint",
  "langchain",
  "langchain-core",
  "langchain-community",
  "langchain-openai",
  "langchain-ollama",
  "langchain-chroma",
  "langchain-google-genai",
  "ddgs",  
  "duckduckgo-search",
  "tabulate",
  "pytz",
  "python-multipart",
  "Pillow",
  "huggingface-hub",
  "torch",
  "transformers",
  "soundfile",
  "kenlm==0.3.0",
  "pyctcdecode @ git+https://github.com/mediacatch/pyctcdecode.git@ff49fc562bf8fc5d6697d4dcd34188dd630cc977"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"